requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
groq>=0.9.0
pandas>=2.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
aiofiles>=23.0.0
